requirements.txt
Django==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
python-decouple==3.8
requests==2.31.0
